+++ releases/translations/push-english-base-translation.sh
@@ -1,0 +1,4 @@
+cd ~/repos/w/wekan-php/public
+tx push -s
+cd ~/repos/w/wekan-php
+