+++ releases/translations/push-translation.sh
@@ -1,0 +1,4 @@
+cd ~/repos/w/wekan-php/public
+tx push -t -l $1
+cd ~/repos/w/wekan-php
+